% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{as.data.frame.ecocbo_data}
\alias{as.data.frame.ecocbo_data}
\title{Coerce ecocbo_data to data.frame}
\usage{
\method{as.data.frame}{ecocbo_data}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{ecocbo_data}.}

\item{row.names}{Passed to \code{as.data.frame()}.}

\item{optional}{Passed to \code{as.data.frame()}.}

\item{...}{Additional arguments, ignored.}
}
\value{
A data.frame representation of the \code{Results} component.
}
\description{
Coerce ecocbo_data to data.frame
}
\keyword{internal}
