% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{balanced_sampling2_3}
\alias{balanced_sampling2_3}
\title{Balanced sampling 2_3}
\usage{
balanced_sampling2_3(
  i,
  NN,
  Y1,
  mn,
  nSect,
  M,
  N,
  H0Sim,
  HaSim,
  resultsHa,
  factEnv,
  transformation,
  method,
  model
)
}
\arguments{
\item{i}{pointer to the index in the list of experimental designs to try.}

\item{NN}{Total number of iterations that the experiment will consider.}

\item{Y1}{A data frame with two columns, one indicates the auxiliary variables
on which the sample must be balanced and the other contains the vector of integers
that defines the primary sampling units. This is used by
\code{sampling::balancedtwostage()}}

\item{mn}{A data frame with two columns, one indicates the number of primary
sampling units to be selected and the other the number of second-stage sampling
units to be selected in the iteration. This is used by
\code{sampling::balancedtwostage()}}

\item{nSect}{Total number of sectors to be simulated in each data set.}

\item{M}{Total number of replicates to be simulated in each data set.}

\item{N}{Total number of samples to be simulated in each site.}

\item{H0Sim}{simulated community from \code{SSP::simdata()} in which H0 is true.}

\item{HaSim}{simulated community from \code{SSP::simdata()} in which H0 is false.}

\item{resultsHa}{helper matrix that stores labels and later the results.}

\item{factEnv}{a data frame for indicating the treatment, replicate and sampling
unit lables in each experiment.}

\item{transformation}{Mathematical function to reduce the weight of very
dominant species.}

\item{method}{appropriate distance/dissimilarity metric (e.g. Gower,
Bray–Curtis, Jaccard, etc).}

\item{model}{which algorithm to use for the calculation? At the moment, the only
option is "nested.symmetric".}
}
\value{
a data frame with values for observed F (for H0 and Ha), and the Ha mean
squares for residuals (MS_R) and variation among sites (MS_B(A)).
}
\description{
Develops the experimental design based on the provided conditions
}
\references{
Underwood, A. J. (1997). Experiments in ecology: their logical
design and interpretation using analysis of variance. Cambridge university
press.
}
\seealso{
\code{\link[sampling:balancedtwostage]{sampling::balancedtwostage()}}
}
\author{
Edlin Guerra-Castro (\email{edlinguerra@gmail.com}), Arturo Sanchez-Porras
}
\keyword{internal}
