\name{ecocbo-package}

\alias{ecocbo-package}
\alias{ecocbo}

\docType{package}

\title{
Cost-benefit optimization in ecological sampling from simulated data
}

\description{

}

\details{

}

\references{
Underwood, A. J., Underwood, A. J., & Wnderwood, A. J. (1997). Experiments in ecology: their logical design and interpretation using analysis of variance. Cambridge university press.

Underwood, A. J., & Chapman, M. G. (2003). Power, precaution, Type II error and sampling design in assessment of environmental impacts. Journal of Experimental Marine Biology and Ecology, 296(1), 49-70.

Anderson, M. J. (2014). Permutational multivariate analysis of variance (PERMANOVA). Wiley statsref: statistics reference online, 1-15.

Guerra‐Castro, E. J., Cajas, J. C., Simões, N., Cruz‐Motta, J.J., & Mascaró, M. (2021). SSP: an R package to estimate sampling effort in studies of ecological communities. Ecography, 44(4), 561-573.
}

\author{
The \pkg{ecocbo} development team is Edlin Guerra-Castro and Arturo Sanchez-Porras.
}

\examples{

}

\keyword{ package}
