% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epiDat}
\alias{epiDat}
\title{Dataset on species count of marine communities.}
\format{
A data frame with count of individuals for 24 observations on 151 species.
}
\source{
Data available from the Dryad Digital Repository:
\doi{10.5061/dryad.3bk3j9kj5} (Guerra-Castro et al. 2020).
}
\usage{
epiDat
}
\description{
This is a dataset containing a subset from the epibionts dataset from \code{SSP}
which was made by using the three local communities that differ the most.
}
\keyword{datasets}
