% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macrofDat}
\alias{macrofDat}
\title{Dataset on species count of coastal macrofauna.}
\format{
A dataframe with counts of individuals for 43 observations on 34 species.
}
\source{
Data available from the GitHub Digital Repository:
\url{https://github.com/edlinguerra/IA206320_publico/tree/main/datos} (Guerra-Castro
et al. 2022).
}
\usage{
macrofDat
}
\description{
This is a dataset containing a subset from the macrofauna recorded in the
PAPIIT experiment.
}
\keyword{datasets}
