% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova_reduced.R
\name{permanova_reduced}
\alias{permanova_reduced}
\title{Calculate pseudoF by using PERMANOVA}
\usage{
permanova_reduced(x, factor, type = "P", method = "bray")
}
\arguments{
\item{x}{Ecological community to be used.}

\item{factor}{List of labels for the sites.}

\item{type}{PERMANOVA algorithm to be used. Options are "P" and "BF". "P" will also throw out the values for MeanSquares, as they are used for other functions in the package.}

\item{method}{Dissimilaty index to be used to calculate dissimilarity matrix.}
}
\value{
A data frame with information regarding pseudoF and, if required, mean square values.
}
\description{
Calculate pseudoF by using PERMANOVA
}
\examples{
perHa1 <- epiDat[,-1]
envHa1 <- epiDat[,1]

permanova_reduced(x = perHa1, factor = envHa1, type = "P", method = "bray")
permanova_reduced(x = perHa1, factor = envHa1, type = "BF", method = "bray")
}
