% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cbo.R
\name{print.cbo_result}
\alias{print.cbo_result}
\title{S3Methods for Printing}
\usage{
\method{print}{cbo_result}(x, ...)
}
\arguments{
\item{x}{Object from \code{ecocbo::sim_cbo()} function.}

\item{...}{Additional arguments}
}
\value{
Prints a summary for the results of \code{ecocbo::sim_cbo()} function,
showing in an ordered matrix the suggested experimental design, according to
cost and estimated power.
}
\description{
prints for \code{ecocbo::sim_cbo()} objects.
}
\keyword{internal}
