% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{print.ecocbo_data}
\alias{print.ecocbo_data}
\title{Print method for ecocbo_data}
\usage{
\method{print}{ecocbo_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ecocbo_data}.}

\item{...}{Additional arguments, ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print method for ecocbo_data
}
\keyword{internal}
