% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_es.R
\name{print.effect_size_data}
\alias{print.effect_size_data}
\title{Print method for effect_size_data objects}
\usage{
\method{print}{effect_size_data}(x, digits = max(3L, getOption("digits") - 2L), preview_n = 6L, ...)
}
\arguments{
\item{x}{An object of class \code{effect_size_data}.}

\item{digits}{Number of digits to print.}

\item{preview_n}{Number of rows to preview.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for effect_size_data objects
}
