% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_power.R
\name{print.pilot_rarefaction_workflow}
\alias{print.pilot_rarefaction_workflow}
\title{Print a pilot rarefaction workflow object}
\usage{
\method{print}{pilot_rarefaction_workflow}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"pilot_rarefaction_workflow"}, usually returned
by \code{\link[=empirical_power]{empirical_power()}}.}

\item{...}{Additional arguments passed to or from other methods. Currently
ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a compact summary of an object returned by \code{\link[=empirical_power]{empirical_power()}}.
}
