% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scompvar.R
\name{scompvar}
\alias{scompvar}
\title{Simulated components of variation}
\usage{
scompvar(data, m = NULL, n = NULL)
}
\arguments{
\item{data}{List that results from beta().}

\item{m}{Site label to be used as basis for the computation.}

\item{n}{Number of samples to be considered.}
}
\value{
A data frame containing the values for the variation component among sites and in the residuals.
}
\description{
Simulated components of variation
}
\examples{
scompvar(data = epiBetaR)
scompvar(data = epiBetaR, m = 2, n = 5)
}
