% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_beta.R
\name{sim_beta}
\alias{sim_beta}
\title{Calculate beta and power out of simulated samples}
\usage{
sim_beta(simH0, simHa, n, m, k = 50, alpha = 0.05)
}
\arguments{
\item{simH0}{Simulated community from SSP::simdata in which H0 is true.}

\item{simHa}{Simulated community from SSP::simdata in which H0 is false.}

\item{n}{Maximum number of samples to consider.}

\item{m}{Maximum number of sites.}

\item{k}{Number of resamples the process will take. Defaults to 50.}

\item{alpha}{Level of significance. Defaults to 5\%.}
}
\value{
A list with two data frames: a data frame containing the values of beta for different sampling efforts, and a data frame containing the results of the sampling.
}
\description{
Calculate beta and power out of simulated samples
}
\examples{
# Load data and adjust it.
epiH0 <- epiDat
epiH0[,"site"] <- as.factor("T0")
epiHa <- epiDat
epiHa[,"site"] <- as.factor(epiHa[,"site"])

# Calculate simulation parameters.
parH0 <- SSP::assempar(data = epiH0, type = "counts", Sest.method = "average")
parHa <- SSP::assempar(data = epiHa, type = "counts", Sest.method = "average")

# Simulation.
simH0 <- SSP::simdata(parH0, cases = 3, N = 1000, sites = 1)
simHa <- SSP::simdata(parHa, cases = 3, N = 100, sites = 10)

sim_beta(simH0, simHa, n = 10, m = 3, k = 20, alpha = 0.05)
}
