% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cbo.R
\name{sim_cbo}
\alias{sim_cbo}
\alias{simcbo}
\title{Simulated Cost-Benefit Optimization}
\usage{
sim_cbo(data, cn, cm = NULL)
}
\arguments{
\item{data}{Object of class \code{"ecocbo_beta"} obtained from \code{\link[=sim_beta]{sim_beta()}}.}

\item{cn}{Numeric. Cost per sampling unit.}

\item{cm}{Numeric. Cost per replicate.}
}
\value{
A data frame containing the optimized values for \code{m} number of
sites to sample and \code{n} number of samples per site.
}
\description{
Applies a cost-benefit optimization based on a desired level of statistical
power and the sampling cost.
}
\examples{
compVar <- scompvar(data = simResults)

# Optimization of single factor experiment
sim_cbo(data = epiBetaR, cn = 80)

}
\references{
\itemize{
\item Underwood, A. J. (1997). Experiments in ecology: their logical
design and interpretation using analysis of variance. Cambridge university
press.
\item Underwood, A. J., & Chapman, M. G. (2003). Power, precaution,
Type II error and sampling design in assessment of environmental impacts.
Journal of Experimental Marine Biology and Ecology, 296(1), 49-70.
}
}
\seealso{
\code{\link[=sim_beta]{sim_beta()}}
\code{\link[=plot_power]{plot_power()}}
\code{\link[=scompvar]{scompvar()}}
\code{\link[=Underwood_cbo]{Underwood_cbo()}}
}
\author{
Edlin Guerra-Castro (\email{edlinguerra@gmail.com}), Arturo Sanchez-Porras
}
