\name{simHaDat}
\alias{simHaDat}
\docType{data}
\title{
Dataset for the example of `sim_beta`
}
\description{
This dataset contains half of the necessary data to test \code{sim_beta}, the other half being \code{\link{simH0Dat}}. These datasets are the result of applying the preparation steps mentioned in the vignette "ecocbo-guide": (1) subset `epibionts`, (2) adjust the `site` labels, (3) apply \code{assempar}, and (4) apply \code{simdata}.
}
\usage{data("simHaDat")}
\format{
  A list with 3 data frames with dimensions 1000 x 95.
  List of 3
 $ :'data.frame':	1000 obs. of  95 variables:
  ..$ X1   : num [1:1000] 0 0 0 0 0 0 1 0 0 0 ...
  ..$ X2   : num [1:1000] 0 0 0 0 0 18 0 0 0 0 ...
  ..$ X3   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X4   : num [1:1000] 0 0 1 0 0 0 0 0 0 0 ...
  ..$ X5   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X6   : num [1:1000] 0 7 0 0 0 0 0 0 0 0 ...
  ...
  ..$ N    : int [1:1000] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ sites: Factor w/ 1 level "1": 1 1 1 1 1 1 1 1 1 1 ...
 $ :'data.frame':	1000 obs. of  95 variables:
  ..$ X1   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X2   : num [1:1000] 0 4 9 0 0 0 0 0 0 0 ...
  ..$ X3   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X4   : num [1:1000] 0 1 0 0 0 0 0 0 0 0 ...
  ..$ X5   : num [1:1000] 43 0 0 0 0 0 0 26 0 0 ...
  ..$ X6   : num [1:1000] 0 7 0 0 7 0 0 0 36 0 ...
  ...
  ..$ N    : int [1:1000] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ sites: Factor w/ 1 level "1": 1 1 1 1 1 1 1 1 1 1 ...
 $ :'data.frame':	1000 obs. of  95 variables:
  ..$ X1   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X2   : num [1:1000] 0 0 0 0 12 0 0 0 0 0 ...
  ..$ X3   : num [1:1000] 0 0 24 0 0 0 0 0 0 0 ...
  ..$ X4   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X5   : num [1:1000] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ X6   : num [1:1000] 11 6 0 0 0 5 0 0 0 0 ...
  ...
  ..$ N    : int [1:1000] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ sites: Factor w/ 1 level "1": 1 1 1 1 1 1 1 1 1 1 ...
  }
}
\source{
Data available from the Dryad Digital Repository: <http://dx.doi.org/10.5061/dryad.3bk3j9kj5> (Guerra-Castro et al. 2020).
}
\references{
Guerra-Castro, E. J. et al. 2016. Scales of spatial variation in tropical benthic assemblages and their ecological relevance: epibionts on Caribbean mangrove roots as a model system. – Mar. Ecol. Prog. Ser. 548: 97–110.
}
\examples{
data(simHaDat)
sim_beta(simHaDat, simHaDat, n = 10, m = 3, k = 20, alpha = 0.05)
}
\keyword{datasets}
