% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simResultsNested}
\alias{simResultsNested}
\title{Data set containing the results of applying ecocbo::prep_data() to a nested
factors experiment.}
\format{
An object of class "ecocbo_data", also a list containing one data frame.
The format is:
\describe{
\item{$Results}{
\describe{
\item{dat.sim}{simulation from which the results are obtained.}
\item{k}{number of resample for the result.}
\item{m}{number of sites considered for the result.}
\item{n}{number of replicates within each site for the result.}
\item{pseudoFH0}{observed F value for the experimental design, when all observations belong to one site.}
\item{pseudoFHa}{observed F value for the experimental design, when observations belong to different sites.}
\item{MSB(A)}{calculated mean squares among sites in the experiment.}
\item{MSR}{calculated mean squares for the residuals in the experiment.}
}}
\item{$model}{"single.factor"}
\item{attribute}{class: ecocbo_data}
}
}
\source{
Source data is available from
\url{https://github.com/edlinguerra/IA206320_publico/tree/main/datos}
(Guerra-Castro et al. 2020).
}
\usage{
simResultsNested
}
\description{
The dataset contains the results of applying ecocbo::prep_data() to epiDat.
The result is a list with one level: $Results is a data frame with the results
of applying PERMANOVA to epiDat a number of times, it contains the values of
pseudoF and the mean squares for different repeated sampling efforts.
}
\details{
This dataset can be used to study the variability of the pseudoF-statistic,
beta and the power when an experiment is applied to a varying number of samples,
sampling units, or sampling sites.
}
\keyword{datasets}
