% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SS}
\alias{SS}
\title{Sum of squares using Huygen Theorem}
\usage{
SS(d)
}
\arguments{
\item{d}{distance matrix from which the sum of squares will be calculated.}
}
\value{
A numeric vector containing the dimension for the distance matrix, and
the value for the sum of squares for the matrix.
}
\description{
Calculates sum of squares using Huygen theorem as implemented by Anderson (2014).
}
\references{
Anderson, M. J. (2014). Permutational multivariate analysis of
variance (PERMANOVA). Wiley statsref: statistics reference online, 1-15.
}
\author{
Edlin Guerra-Castro (\email{edlinguerra@gmail.com}), Arturo Sanchez-Porras
}
\keyword{internal}
