% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{summary.ecocbo_data}
\alias{summary.ecocbo_data}
\title{Summary method for ecocbo_data}
\usage{
\method{summary}{ecocbo_data}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ecocbo_data}.}

\item{...}{Additional arguments, ignored.}
}
\value{
A summary list of the main object components.
}
\description{
Summary method for ecocbo_data
}
\keyword{internal}
