% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_es.R
\name{validate_effect_size_data}
\alias{validate_effect_size_data}
\title{Validate an effect_size_data object}
\usage{
validate_effect_size_data(x)
}
\arguments{
\item{x}{A data frame or \code{effect_size_data} object.}
}
\value{
Invisibly returns \code{TRUE} if validation passes.
}
\description{
Validate an effect_size_data object
}
\keyword{internal}
