\name{epiBetaR}
\alias{epiBetaR}
\docType{data}
\title{
Simulated results for analizing statistical power with ecocbo::sim_beta
}
\description{
The dataset contains the results of applying ecocbo::sim_beta to epiDat. The result is a list with two levels. $Power A data frame with the results of statistical power according to different sampling efforts. $Results A data frame with the results of applying PERMANOVA to epiDat a number of times, it contains the values of pseudoF and the mean squares for different repeated sampling efforts. This dataset can be used to study the variability of the pseudoF-statistic, beta and the power when an experiment is applied to a varying number of samples, sampling units, or sampling sites.
}
\usage{data("epiBetaR")}
\format{
  The format is:
List of 2
 $ Power  :'data.frame':	18 obs. of  7 variables:
  ..$ m    : num [1:18] 2 2 2 2 2 2 2 2 2 3 ...
  ..$ n    : num [1:18] 2 3 4 5 6 7 8 9 10 2 ...
  ..$ Power: num [1:18] 0.333 0.567 0.65 0.817 0.95 ...
  ..$ Beta : num [1:18] 0.667 0.433 0.35 0.183 0.05 ...
  ..$ fCrit: num [1:18] 1.64 1.58 1.66 1.69 1.45 ...
  ..$ AMSHa: num [1:18] 0.338 0.336 0.34 0.339 0.332 ...
  ..$ RMSHa: num [1:18] 0.47 0.557 0.658 0.722 0.769 ...
 $ Results:'data.frame':	1080 obs. of  8 variables:
  ..$ dat.sim  : num [1:1080] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ k        : num [1:1080] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ m        : num [1:1080] 2 2 2 2 2 2 2 2 2 2 ...
  ..$ n        : num [1:1080] 2 2 2 2 2 2 2 2 2 2 ...
  ..$ pseudoFH0: num [1:1080] 1.078 1.701 1.055 0.786 0.976 ...
  ..$ pseudoFHa: num [1:1080] 0.911 1.45 1.353 0.91 1.102 ...
  ..$ AMSHa    : num [1:1080] 0.356 0.381 0.509 0.346 0.424 ...
  ..$ RMSHa    : num [1:1080] 0.391 0.263 0.377 0.381 0.384 ...
}
\examples{
data(epiBetaR)
ecocbo::plot_power(epiBetaR, m = 2, n = 7, method = "both")
ecocbo::scompvar(data = epiBetaR)
}
\keyword{datasets}
