\name{epiBetaR}
\alias{epiBetaR}
\docType{data}
\title{
Dataset containing an example of the results of applying `ecocbo::sim_beta`.
}
\description{
The dataset contains the results of applying ecocbo::sim_beta to epiDat. The result is a list with two levels. $Power A data frame with the results of statistical power according to different sampling efforts. $Results A data frame with the results of applying PERMANOVA to epiDat a number of times, it contains the values of pseudoF and the mean squares for different repeated sampling efforts. This dataset can be used to study the variability of the pseudoF-statistic, beta and the power when an experiment is applied to a varying number of samples, sampling units, or sampling sites.
}
\usage{data("epiBetaR")}
\format{
  The format is:
List of 2
 $ Power  :'data.frame':	12 obs. of  7 variables:
  ..$ m    : num [1:12] 2 2 2 2 3 3 3 3 4 4 ...
  ..$ n    : num [1:12] 2 3 4 5 2 3 4 5 2 3 ...
  ..$ Power: num [1:12] 0.233 0.517 0.7 0.817 0.333 ...
  ..$ Beta : num [1:12] 0.767 0.483 0.3 0.183 0.667 ...
  ..$ fCrit: num [1:12] 1.74 1.56 1.63 1.61 1.7 ...
  ..$ AMSHa: num [1:12] 0.328 0.335 0.325 0.334 0.322 ...
  ..$ RMSHa: num [1:12] 0.455 0.554 0.651 0.736 0.47 ...
 $ Results:'data.frame':	720 obs. of  8 variables:
  ..$ dat.sim  : num [1:720] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ k        : num [1:720] 1 2 3 4 5 6 7 8 9 10 ...
  ..$ m        : num [1:720] 2 2 2 2 2 2 2 2 2 2 ...
  ..$ n        : num [1:720] 2 2 2 2 2 2 2 2 2 2 ...
  ..$ pseudoFH0: num [1:720] 1.027 0.783 1.222 0.519 1.048 ...
  ..$ pseudoFHa: num [1:720] 1.75 1.15 1.68 0.97 1.77 ...
  ..$ AMSHa    : num [1:720] 0.486 0.268 0.472 0.418 0.557 ...
  ..$ RMSHa    : num [1:720] 0.278 0.233 0.281 0.431 0.314 ...
 - attr(*, "class")= chr "ecocbo_beta"
}
\details{
This dataset comes from applying `ecocbo::sim_beta()` to the basic data from `ecocbo::epiDat`
}
\source{
Data available from the Dryad Digital Repository: <http://dx.doi.org/10.5061/dryad.3bk3j9kj5> (Guerra-Castro et al. 2020).
}
\references{
Guerra-Castro, E. J. et al. 2016. Scales of spatial variation in tropical benthic assemblages and their ecological relevance: epibionts on Caribbean mangrove roots as a model system. – Mar. Ecol. Prog. Ser. 548: 97–110.
}
\examples{
data(epiBetaR)
plot_power(epiBetaR, n = 3, m = 4, method = "both")
}
\keyword{datasets}
