\name{epiDat}
\alias{epiDat}
\docType{data}
\title{Dataset on species count of marine communities}
\description{This is a dataset containing a subset from the epibionts dataset from `SSP` which was made by using the three local communities that differ the most.}
\usage{data("epiDat")}
\format{
  A data frame with count of individuals for 24 observations on 151 species.
}
\source{
Data available from the Dryad Digital Repository: <http://dx.doi.org/10.5061/dryad.3bk3j9kj5> (Guerra-Castro et al. 2020).
}
\references{
Guerra-Castro, E. J. et al. 2016. Scales of spatial variation in tropical benthic assemblages and their ecological relevance: epibionts on Caribbean mangrove roots as a model system. – Mar. Ecol. Prog. Ser. 548: 97–110.
}
\examples{
data("epiDat")

str(epiDat)
}
\keyword{datasets}
