% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_es.R
\name{new_effect_size_data}
\alias{new_effect_size_data}
\title{Create an effect_size_data object}
\usage{
new_effect_size_data(data, centroid_distances = NULL, pcoa = NULL, call = NULL)
}
\arguments{
\item{data}{A data frame with at least \code{reduction_level}
and \code{ecological_effect}. Inferential columns such as
\code{omega2}, \code{R2}, or \code{pseudoF} are optional but recommended.}

\item{centroid_distances}{Optional list or object storing
centroid distance matrices.}

\item{pcoa}{Optional list or object storing PCoA results.}

\item{call}{Optional matched call.}
}
\value{
An object of class \code{effect_size_data}.
}
\description{
Create an effect_size_data object
}
