% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova_oneway.R
\name{permanova_oneway}
\alias{permanova_oneway}
\title{Calculate pseudoF by using PERMANOVA}
\usage{
permanova_oneway(x, factor, type = "P", method = "bray")
}
\arguments{
\item{x}{Ecological community to be used.}

\item{factor}{List of labels for the sites.}

\item{type}{PERMANOVA algorithm to be used. Options are "P" and "BF". "P" will also throw out the values for MeanSquares, as they are used for other functions in the package.}

\item{method}{Dissimilaty index to be used to calculate dissimilarity matrix.}
}
\value{
A data frame with information regarding pseudoF and, if required,
mean square values.
}
\description{
\code{permanova_oneway} performs a "permutational multivariate analysis
of variance", which can be used to asses the significance of differences
in community structure between groups.
}
\examples{
perHa1 <- epiDat[,-1]
envHa1 <- epiDat[,1]

permanova_oneway(x = perHa1, factor = envHa1, type = "P", method = "bray")
permanova_oneway(x = perHa1, factor = envHa1, type = "BF", method = "bray")
}
\references{
Anderson, M. J. (2014). Permutational multivariate analysis of
variance (PERMANOVA). Wiley statsref: statistics reference online, 1-15.
}
\author{
Edlin Guerra-Castro (edlinguerra@gmail.com), Arturo Sanchez-Porras
}
