% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_es.R
\name{plot.effect_size_data}
\alias{plot.effect_size_data}
\title{Plot effect_size_data}
\usage{
\method{plot}{effect_size_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{effect_size_data}.}

\item{...}{Additional arguments, ignored.}
}
\value{
A ggplot representation of the \code{effect_size_data} object.
}
\description{
Plot effect_size_data
}
