% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power.R
\name{plot_power}
\alias{plot_power}
\title{Power curves for different sampling efforts}
\usage{
plot_power(data, n = NULL, m, method = "both")
}
\arguments{
\item{data}{List that results from beta().}

\item{n}{Defaults to NULL, and then the function computes the number of samples (n) that results in a sampling effort close to 95\% in power. If provided, said number of samples will be used.}

\item{m}{Site label to be used as basis for the plot.}

\item{method}{The desired plot. Options are "power", "density" or "both". "power" plots the power curve, "density" plots the density distribution of pseudoF, and "both" draws both plots one next to the other.}
}
\value{
A plot
}
\description{
Power curves for different sampling efforts
}
\examples{
plot_power(data = epiBetaR, n = 4, m = 2, method = "both")
plot_power(data = epiBetaR, n = NULL, m = 3, method = "both")
}
