% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power.R
\name{plot_power}
\alias{plot_power}
\title{Power curves for different sampling efforts}
\usage{
plot_power(data, m, n = NULL, method = "both")
}
\arguments{
\item{data}{List that results from beta().}

\item{m}{Site label to be used as basis for the plot.}

\item{n}{Defaults to NULL, and then the function computes the number of samples (n) that results in a sampling effort close to 95\% in power. If provided, said number of samples will be used.}

\item{method}{The desired plot. Options are "power", "density" or "both". "power" plots the power curve, "density" plots the density distribution of pseudoF, and "both" draws both plots one next to the other.}
}
\value{
A plot
}
\description{
Power curves for different sampling efforts
}
\examples{
library("SSP")
# Load data and adjust it.
epiH0 <- epiDat
epiH0[,"site"] <- as.factor("T0")
epiHa <- epiDat
epiHa[,"site"] <- as.factor(epiHa[,"site"])

# Calculate simulation parameters.
parH0 <- SSP::assempar(data = epiH0, type = "counts", Sest.method = "average")
parHa <- SSP::assempar(data = epiHa, type = "counts", Sest.method = "average")

# Simulation.
simH0 <- SSP::simdata(parH0, cases = 3, N = 1000, sites = 1)
simHa <- SSP::simdata(parHa, cases = 3, N = 100, sites = 10)

betaResults <- beta(simH0, simHa, n = 10, m = 3, k = 50, alpha = 0.05)

plot_power(data = betaResults, m = 2, n = 7, method = "both")
plot_power(data = betaResults, m = 3, n = NULL, method = "both")
}
