% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_beta.R
\name{prints}
\alias{prints}
\alias{print.ecocbo_beta}
\title{S3Methods for Printing}
\usage{
\method{print}{ecocbo_beta}(x, ...)
}
\arguments{
\item{x}{Object from \code{ecocbo::sim_beta} package}

\item{...}{Additional arguments}
}
\value{
Prints \code{ecocbo::sim_beta} object
}
\description{
Prints for \code{ecocbo::sim_beta} objects
}
