% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_beta.R
\name{prints}
\alias{prints}
\alias{print.ecocbo_beta}
\title{S3Methods for Printing}
\usage{
\method{print}{ecocbo_beta}(x, ...)
}
\arguments{
\item{x}{Object from \code{ecocbo::sim_beta()} function.}

\item{...}{Additional arguments}
}
\value{
Prints the result of \code{ecocbo::sim_beta()} function, showing in an
ordered matrix the estimated power for the different experimental designs
that were considered.
}
\description{
prints for \code{ecocbo::sim_beta()} objects.
}
