% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scompvar.R
\name{scompvar}
\alias{scompvar}
\title{Simulated components of variation}
\usage{
scompvar(data, m = NULL, n = NULL)
}
\arguments{
\item{data}{List that results from beta().}

\item{m}{Site label to be used as basis for the computation.}

\item{n}{Number of samples to be considered.}
}
\value{
A data frame containing the values for the variation component among sites and in the residuals.
}
\description{
Simulated components of variation
}
\examples{
library("SSP")
# Load data and adjust it.
epiH0 <- epiDat
epiH0[,"site"] <- as.factor("T0")
epiHa <- epiDat
epiHa[,"site"] <- as.factor(epiHa[,"site"])

# Calculate simulation parameters.
parH0 <- SSP::assempar(data = epiH0, type = "counts", Sest.method = "average")
parHa <- SSP::assempar(data = epiHa, type = "counts", Sest.method = "average")

# Simulation.
simH0 <- SSP::simdata(parH0, cases = 3, N = 1000, sites = 1)
simHa <- SSP::simdata(parHa, cases = 3, N = 100, sites = 10)

betaResults <- beta(simH0, simHa, n = 10, m = 3, k = 50, alpha = 0.05)

scompvar(data = betaResults)
scompvar(data = betaResults, m = 2, n = 5)
}
