% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcbo.R
\name{simcbo}
\alias{simcbo}
\title{Simulated cost-benefit optimization}
\usage{
simcbo(comp.var, multSE, ct, ck, cj)
}
\arguments{
\item{comp.var}{Data frame from scompvar().}

\item{multSE}{Optional. Required multivariate standard error for the sampling experiment.}

\item{ct}{Optional. Total cost for the sampling experiment.}

\item{ck}{Cost per replicate.}

\item{cj}{Cost per unit.}
}
\value{
A data frame containing the optimized values for b (number of sites) and n (number of samples) to consider.
}
\description{
Simulated cost-benefit optimization
}
\examples{
compVar <- scompvar(data = epiBetaR)

simcbo(comp.var = compVar, multSE = NULL, ct = 600, ck = 11, cj = 108)
simcbo(comp.var = compVar, multSE = 0.15, ct = NULL, ck = 11, cj = 108)
}
